feat(layout): add Header in all application pages and create space to the player component
It was moved the Header component to the _app component in order to the Header to be used in all
application pages. Plus, the component Player received the first layers of style and it was placed
as well in the _app component together with the Header component, composing the global layout for
the whole application screen.