feat(home): make the home page consume our server data using NextJS SSG(Static Site Generation)
The home page was initially configured to consume our backend server data using SSG(Static Site
Generation). Meaning that when we access our application it will be more performatic and from time
to time, specifically every 8 hours, it will a proccess called revalidation and our app will make a
api call to our backend. And not final but I had also add a eslint rule for enabling the use of
console.log.