feat: format backend data right after the request is answered and it will utils folder
The duration property of the episode object it will be formatted with the helper function called
'convertDurationToTimeString' so it cant be used in the progress bar of the Player component. Also
the published_at property was renamed to the camelCase format, so it was  written as 'publishedAt'.
Plus, the publishedAt property got formatted by two functions of the 'date-fns' package: format and
parseISO, in order to the result format be '9 abr 20' (e.g).